Consolidate READMEs into Markdown version
Since Markdown is meant to be perfectly readably plain text, it makes
sense to keep only a Markdown version (github dialect). Plain text
consumers can read it just as easily.

The consolidation of the READMEs requires changes in the MANIFEST
and the boilerplate test. The README remains as a symlink.

Add a TODO file to keep track of (internal) change requests.