New method 'list' for listing selected filters
This method necessicates that the object retains not only the code
of the filters, but the names as well.

An alternative implementation would be to store *only* the names,
and then to look up the code on each process() call. This would
also cause the behaviour of an object to change if the list of
registered filters changes after object creation and filling.
It is not clear which behaviour is the desired one.