Replaced 'use feature "unicode_strings"' et al
The pragma 'use feature "unicode_strings"', as well as the 'u'
modifiers on regexes, demand a quite recent version of Perl as
of this writing. This is not ideal, and not even required.

Instead, all text to be filtered is passed through 'utf8::upgrade'
before parsing. This also enables character string semantics, and
is a reasonable change to make.