More TODO: stashes are out, unicode filters in
The "stash" mechanism, along with 'rewind' and 'eof' methods,
came out of the following use case:

The 'whitespace' filter chomps all whitespace at the beginning
*and end* of a line. The line following it is treated the same.
If the lines are printed out again, there is *no* whitespace
between the last word of the first line and the first word of the
second line!

The "stash" mechanism was supposed to give filters a "state", so
'whitespace' could decide whether it had already "printed a line"
and prepend a space to the next line in this case.

This state, however, seriously violates the rule of least surprise.
Unless it's really, really needed, I'll scratch the idea. I'll
modify 'whitespace' to work statelessly with multiple lines instead.