Filter 'whitespace' leaves a space at the end
In order to facilitate joining multiple filtered lines, the
'whitespace' filter no longer removes all whitespace from the end
of a string, but collapses it to a single space instead.

The 'whitespace_nl' filter also mirrors this behaviour, leaving
a single space at texts that end in whitespace which does not
contain a newline character.