Switches 'paragraph' and line-based presets
The 'paragraph' command line switch sets the input to "paragraph
mode" (sets $/ to ''). This means blocks of text are separated
by blank lines, not newline characters.

Most useful with the 'whitespace_nl' filter, or with the new
presets '2l' and '3l', which work like '2' and '3', but with the
'whitespace_nl' filter instead.